requirements.txt
openai>=1.6.1
bitsandbytes
fastapi>=0.104.1
python-dotenv==1.0.0
uvicorn
tenacity
sentencepiece
tiktoken
loguru
transformers>=4.37.0
peft>=0.4.0
accelerate>=0.20.3
sentence-transformers>=2.7.0
transformers_stream_generator
cpm_kernels
einops
sse-starlette>=1.6.1
starlette-context>=0.3.6,<0.4
langchain>=0.1.16

# for rag
langchain_community
nltk
rapidocr_onnxruntime
unstructured[all-docs]
openparse
python-multipart
